# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: eSIM reseller platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/profile:
    get:
      operationId: getProfile
      tags: [user]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateProfile
      tags: [user]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [uid, email, displayName, travelFrequency, referralCode, createdAt]
      properties:
        uid:
          type: string
        email:
          type: string
        displayName:
          type: string
        photoUrl:
          type: ["string", "null"]
        travelFrequency:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        referralCode:
          type: string
        createdAt:
          type: string
    ProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
        travelFrequency:
          type: string
...